refactor(debug): move serialization helpers to separate TU
expr_types is already a "busy" header.
I would like to avoid adding more unnecessary stuff to this header.
